+++ backend/app/models/story.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Story struct {
-	ID         uint      `json:"id" bun:",pk,autoincrement"`
-	HeadlineID uint      `json:"headline_id" bun:",unique,notnull"`
+	ID         int64     `json:"id" bun:",pk,autoincrement"`
+	HeadlineID int64     `json:"headline_id" bun:",unique,notnull"`
 	Headline   *Headline `json:"-" bun:"rel:belongs-to,join:headline_id=id"` // cascade
 	Summary    string    `json:"summary" bun:",type:text"`
 	Content    string    `json:"content" bun:",type:text"`
