+++ backend/app/models/topic.go
@@ -1,7 +1,7 @@
 package models
 
 type Topic struct {
-	ID      uint      `json:"id" bun:",pk,autoincrement"`
+	ID      int64     `json:"id" bun:",pk,autoincrement"`
 	Name    string    `json:"name"`
 	Sources []*Source `json:"-" bun:"rel:has-many,join:id=topic_id"`
 }
